+++ hutool-captcha/src/main/java/cn/hutool/captcha/AbstractCaptcha.java
@@ -70,7 +70,7 @@ public abstract class AbstractCaptcha implements ICaptcha {
 		this.generator = generator;
 		this.interfereCount = interfereCount;
 		// 字体高度设为验证码高度-2，留边距
-		this.font = new Font("Courier", Font.PLAIN, this.height - 2);
+		this.font = new Font("Courier", Font.PLAIN, (int)(this.height * 0.75));
 		createCode();
 	}