+++ hutool-extra/src/test/java/cn/hutool/extra/mail/MailAccountTest.java
@@ -13,6 +13,9 @@ public class MailAccountTest {
 	@Test
 	public void parseSettingTest() {
 		MailAccount account = GlobalMailAccount.INSTANCE.getAccount();
+		account.getSmtpProps();
+		
 		Assert.assertNotNull(account.getCharset());
+		Assert.assertTrue(account.isSslEnable());
 	}
 }